PAC to main tabs nav
I moved the PAC tab down to the tab menu. It just made sense. The
candidate selector works for the most part. The airframes look funny
but that won’t matter in the end product. Chris will add code that will
clear all that up. Just need to add code for the overlay boxes.